Revert "Unecessary fileno(), just do as cf_file_input()"
This reverts commit 44fc841929934a55f690422ee1d25c3190c5c900.

stat() is not the same as fstat().  The first one checks a
file which has the same name as the one we opened.  The second
checks the file which we just opened.

The cf_file_input() checks just check if the filename for
security (globally writeable, etc.).  cf_file_open() will
open the file and read it's contents