Don't reconnect connections, destroy the old one and get a new one
The old reconnect code didn't release the mutex whilst spawning a new connection, presumably to keep the pool in a consistent state.

This would have weird and annoying side effects. It's simpler just to close the connection and either get another existing one or spawn a new one.

The reduction in complexity is worth the small performance penalty of allocating a new connection structure, and linking it into the pool.